Scrolls of flood fully rehydrate grung.
As long as they are not read while confused - which will do the opposite!

This gives this scroll some nice utility and gives grung an out if they don't want to jump in a deep pool. It also might be a lifesaver in places like Gehennom or the plane of fire where the grung might have a constantly dehydrated.