Fix: Grung hydration could reach 0 after a forced dehydrating action.
I had set this up so the death check was in the standard turn checks, but if a player was reduced to 0 via fire attack then it's possible the player could see more action and trigger this sanity check.

I simply moved the death check to be whenever the hydration counter is changed.