New monster: star vampire (from SLASH'EM and dNetHack).
The star vampire as it was in SLASH'EM was always neat, but I felt it's implementation was a bit flawed. After peeking at dNetHack's version, I thought it fit better because slashem's star vampire is implemented as humanoid and enables the player to poly into it and use it like a normalish form and wear armor etc. I don't think that is very true to the lore and it's also an OP form.

This version pulls in some tweaks from dNetHack but also a couple new changes:
* Star vampires are now in the S_UMBER class and appear as a cyan "U". This is because they are not true vampires and shouldn't be treated as such.
* They are no longer humanoid and have no eyes and no limbs.
* They have a 6d6 claw attack and three 2d6 drain blood attacks.