New monster: gorgon hulk.
A horrifying new monster straight out of the depths of NerfHack.

This caps off the U class as the one of the meanest additions. The gorgon hulk has a stoning gaze. However, this gaze is not as potent as Medusa's as it causes slow stoning instead of instant petrification.

This hulk compares similarly to other hulks, however, to prevent it from being an overpowered polyform, I made it wander which will also cause the player to wander when polyd into one. Enjoy!