New spellbook: waterproofing.
Just like the name says - this is a level 2 Matter spell that will keep your items safe from water and rust damage. I figured this could come in very handy for grung that need the occasional dunk but don't want to bother dropping all their gear. Or it could be very useful for crossing water when you are able to swim. In any case, it's probably very welcome after all the expanded item destruction.

Here's the duration of it's effects:
* Unskilled: 5-15 turns
* Basic: 20-30 turns
* Skilled: 45-55 turns
* Expert: 80-90 turns

In addition, this spellbook and any spellbook with the "oilskin" appearance are immune to water damage.