Fix: Saving and restoring as a grung was costing hydration.
At first I thought this was caused by a long or int type of data error, but it turns out that because I had made polymorphing cost hydration for grung that on restore the game would trigger this on setting your race. I fixed this and also reduced the amount of hydration polymorphing costs.