New armor type: bracers, which also occupy the shield slot (from EvilHack).
This implementation mostly follows EvilHack's, just adapted to 3.7 and with new types of bracers.

Bracers weigh 15 (half as much as a small shield). They allow the use of weapons while worn and don't impede spellcasting or martial arts. However, to balance them out so that shields are not rendered obsolete, they grant 0AC. They still benefit from shield skill, but the rate is only 1AC per skill. So basic = 1AC, skilled = 2AC, etc.

We have a nice variety of bracers to help augment the travelers kit too:
* leather bracers
* bracers of poison resistance
* bracers of sleep resistance
* bracers of cold resistance
* bracers vs shapeshifters
* bracers of unchanging
* bracers vs stone

My main motivation for porting these was to give the grung something else to help them survive, but it turns out that this really helps a lot of roles like monks and even valkyries. After valkyries get to around the mid-game, they might have advanced shield a fair amount, but they would be in a predicament about whether to abandon their shield for two-weaponing or a two-handed weapon. This allows them to carry their shield skill forward without a total sacrifice of the experience they had build up.