New spellbook: flesh to stone.
This introduces a high-level counterpart to stone-to-flesh. It's a level 7 matter spell that will afflict the monster hit slow stoning. Like other slow stoning effects, the monster will have 5 turns to cure itself.

I debated the level, but I think matching this with finger of death makes sense since an interesting chunk of monsters resist it:
* A lot of spellcasters are able to instantly shrug off stoning,
* All elementals resist
* All pudding
* All monsters with petrification attacks
* Rock trolls
* Vlad, Cthulhu, the Riders, Juiblex, all quest nemeses
* Most ghostly and wraiths
* A handful of misc monsters.

To balance it out, the spell will stop at the first monster it hits. However, the range will still increase with skill.

I did not implement any item manipulation yet, and I'm honestly not sure how to approach it. Turning meat rings or meat sticks into rings and wands seems iffy. Would it create a random ring or wand? Would we have to retain a memory of the item's previous identity so it would become it's former type? How would it affect food? There are too many unanswered questions so we'll have to revisit this in the future.