Rename the genocide mechanic to exile and nerf heavily.
This change has nothing to do with ethical or political motivations, I have no problem with the scroll of genocide's name in NetHack. However, I do believe that a renaming is in order because I have wanted to nerf the scroll further and when there is a large difference between two mechanics I think renaming helps players adapt. Since the scroll of very important in NetHack I also don't want to surprise players with totally different behavior.

The new behavior of exile/genocide is thus:
* Permanent monster genocides are no longer allowed.
* An uncursed scroll of exile works to wipe out a single species of monster that is in close proximity to the player.
* A blessed scroll of exile wipes out a single species on the level.
* All other effects are identical to the scroll of genocide in vanilla.
* Genocides from thrones have been replaced with a gift of gold.