Fix: werefoo were interfering with hydration.
This was a very strange discovery, while testing hydration I noticed that it was dramatically decreasing at random points in the game. I played a few sessions where nothing out of the ordinary happened, and then all of a sudden I was severely dehyrated out of nowhere.

I believe I have traced the problem to how were-changing mechanics work. So from the start there is a function called mcalcdistress that runs every turn. This in turn calls m_calcdistress for each monster and makes sure that different timers and mechanics are being calculated correctly.

It turns out that toward the bottom of m_calcdistress, this is always called, unconditionally:

	were_change(mtmp);

If we go further into were_change, it does some pretty standard processing for werefoo, but also - at the bottom of this function, it also calls this unconditionally:

	set_uasmon(); /* new_were() doesn't do this */

When set_uasmon is called, it should usually be acting to polymorph the player but here I'm not sure why it's being called because there's a specific function for player-werefoo changing called you_were().

I think the solution is just removing that set_uasmon call, and I'll submit this to the devteam since it just seems odd.