Allow dehydration to be cured by prayer.
I think this is a reasonable accomodation for grung - we have a similar cure for starving so they can parallel eachother.

If a grung is at or below 250 hydration, we consider them to be 'dehydrated', they are then eligible to pray for some help. This counts as a major trouble, and if granted, their god gives them 1000 hydration. So not a full refill, but enough to hopefully get somewhere wet fast.