Undead Slayers max skills vary dynamically based on their starting kits.
In SLASH'EM, they had 4 different starting kits which I always thought made the role interesting. This extends that idea and limits all the potential kit skills to basic: long sword, short sword, crossbow, spear, and whip. If your starting kit includes one of the weapons in that category, you will be able to reach expert. If you start with the silver short sword, that unlocks skilled potential for long sword.

Slayers can always reach expert in daggers.

I'm going to follow up with a martial arts patch so that the starting whip combo isn't as lame.