New minetown variant: forge town.
This is something I have wanted to do for a while because my wife's parents live in an old blacksmithing town called Wira-Bruk. This town is kind of modeled around that place and features a forging area with two forges, large armor and weapon shops, a candle shop, a tool shop, and a deli. There is an altar but no attending priest. There is also a lack of a general store.

This level may look familiar because it's based on the old Kobold level from SLASH'EM. I find designing maps from scratch quite difficult so that gave me something decent to work from.