Displacement, invisibility, and darkness offer limited protection from spellcasters.
After factoring out the code from gazemu, I was able to use it for this. I had to lower the odds of protection though, otherwise it was too powerful. The odds now match closer to what is going on in monmove with monsters finding you when you are displaced or invisible.