Adjust weapon enchantment.
Being able to reach +13 is very powerful - too powerful some might say... But I like it. However, I think that players should not be able to reach it so quickly. I'm curbing back the enchantment gain on scrolls of enchant weapon to vanilla levels, so that after +5, a scroll of enchant weapon only grants +1. This means that if you are able to get a weapon from +5 to +7 with a blessed scroll, it will still take 6 more scrolls to max out.