Remove magic markers from inventory (again).
"I am altering the deal. Pray I do not alter it any further." - Darth Vader.

All joking aside, I think removing magic markers presents a meaningful and legitimate challenge for players. EvilHack has already demonstrated this effectively. My initial reasoning for guaranteeing magic markers was to discourage start scumming, but in practice, this grants an enormous advantage to any character. With a guaranteed marker, players can rest easy knowing they’ll have access to highly enchanted weapons or armor as soon as they find scrolls of enchant weapon or enchant armor. A single marker, especially when paired with a source of recharging, can generate an abundance of these scrolls.
Additionally, this setup gives Grung an exceptionally reliable way to produce scrolls of flood. While not a game-breaking issue, it does undercut part of the intended challenge of the role. Removing the guaranteed markers ensures that players must work harder for these benefits, making the experience more balanced and rewarding.

I also believe that with the addition of the following markers, having roles start with a marker is just too much:
* Markers are a possible prize for Sokoban
* A marker can be found in the Boulder Bonanza mine's end level.
* A magic marker might generate in a chest in the Lost Tomb
* Monks receive a marker as a crowning gift.