Quest nemeses, demon lords and princes, will perform one warp to get to you and switch to standard movement.
There was a major wrinkle in my porting of covetous non-warper behavior from xNetHack, these monsters would not get to do their standard greeting and the demons didn't seem to be performing their bribe-demands. I think this is a reasonable solution and it makes the initial encounter very dangerous but allows the player to maneuver out in creative ways.