Death drops decrease as games get longer.
For every 5000 turns the player has advanced, the chance of a death drop does down by 1. This is a nerf to slowly starve players out and to reduce inventory bloat toward the end of a game. It's also a good reason to play fast.

The chance of a death drop is traditionally 1 in 6. So, at 5k turns it would decrease to 1 in 7, 10k turns, 1 in 8 and so on.