Only one of the priests in Moloch's Temple tends the altar, the others free-range.
I noticed that after converting the altar as an orc, only one of the priests became angry. I always thought the generation of these priests was a bit odd, but this confirmed that only one of the priests is truly tied to the altar. I also like the idea that having the 3 other priests free makes the level more difficult. Otherwise, a player can back out of the sanctum room whenever the priests get too close.