New monster: elder minotaur (from EvilHack).
This is a very nasty addition. I made a few updates just so it hurts a little more. They now also regenerate and get death resistance, poison resistance, and sleep resistance. Also, in EvilHack they had the M2_WANDER tag for some reason - I removed this and gave them flanking instead. I also updated the tile because it was identical to the minotaur's.

I also followed Evil's placement and added these to the plane of earth, one of Demogorgon's levels, and Geryon's lair.