Turn the Vibrating Square level into a special level.
This is adapted from the xNetHack change. I wanted a more dramatic and symmetrical level for the vibrating square though - and especially something circular.  I found a pretty good map to use as a template in dNetHack so I used that as the base and embellished it a bit.

I had to hard-code the coordinates into create_trap - I'm not satisfied with that long term but it it is a solution for now. If the map bugs out for some reason, it still makes the square randomly (and yes I did test this).

There might be questions as to why I removed the "hunt for the square" aspect - and I think it's something that not many people truly enjoy. I also have added more mandatory puzzles to the Wizard's Tower and portal searching the Planes - so those more than compensate for this.