Reduce random items in hell and demon lairs.
I believe hell has too many items and results in an over-abundance for the endgame. In this commit I reduced the general items for random hell fill levels by about half. For the demon lairs, this was also a good opportunity to make sure most of the demon lairs have similar loot. Some of the newer lairs (especially Lethe versions) had exorbitant caches of loot compared to other versions of the same level. I reduced the default by a bit and applied that to all demon lairs within reason.

I also trimmed down some of the items in the Lost Tomb since it felt a bit excessive.