Refactor and tone down the chest item generation in Moloch's Temple.
There were 9 chests with 3 random objects each, a wax candle, and gold. I think this was too many items so I reduced the chances of the 2nd and 3rd objects. This level is more about the altar anyway. If you slaughter the 4 priests that's also quite a few spellbooks and items.