Remove soko1-7 (duplicate of soko3-3), and add the two Grunt Soko end levels.
I kept getting a weird feeling of deja-vu when playing through soko1-7 and I finally realized it was because the level copied SLASH'EM's level 3c but with the ending zoo and prizes added.

I decided to remove 1-7 because it duplicated the sack-and-rings prize that is already in 1-6 and I felt it should be more rare. I had the two Grunt levels fully ported already and on the backburner, but just hadn't enabled them since it put the level count into double digit territory, but this seemed like a good time to bring them in. They also have the new "standard" selection of 4 random prizes.