New mage spell for monster spellcasters: teleport.
This is a way for higher level spellcasters to really go after the player. After curbing warping from nemesis and demon lords, I still felt like there should be *some* warping, just not as crazy as covetous behavior was in the past. After walking home one day, I realized warping would probably be best if it was a spell and could be adjusted in what monster difficulty it could apply to. It would also be slightly regulated by the spellcasting cooldown that monsters get.

This will require some extensive playtesting to get right.