Force-fighting displacers can also result in displacement.
This addresses the issue I submitted for NetHack: https://github.com/NetHack/NetHack/issues/1377.

This still allows for the 'You attack thin air' message, but if the monster is marked to displace you, you will be displaced even when force-fighting.

I just tested the basics against a displacer beast and a standard monster.