Lord Surtur can cast mage spells.
This is in line with lore, but it also allows him to get out of his fortress if the drawbridge is destroyed. While playtesting, I ran into the frustrating situation where I had no way to get in. If he can cast spells, he can also cast Teleport to get out. It also makes him a nastier nemesis.