Fumbling only causes 50% of kicks to be ineffective.
While implementing the greased feet mechanics, I thought the player would be able to kick off the grease after a while by trying to kick monsters. However, because fumbling renders kicking almost completely useless, this wasn't very likely. I also improved the chance of goop wearing off from 5% to 10%, so that with the fumbling change it should still be effectively 50%.