Create foundation for item alignments.
In this commit I add the ability for items to spawn with the basic three alignments: neutral, chaotic, and lawful. I used EvilHack's recent build quality system as a template and it worked quite easily. With this system I'm hoping to add a bit more spice to the ordinary items that are scattered around the dungeon, as well as complicate the forging process just a bit more. I have a lot of ideas to flesh this mechanic out - should be interesting.