Cross-aligned items can blast you.
This is a rough first draft but we do have a working proof of concept. I was able to hi-jack touch_artifact for my purposes and get it also to work for aligned items. Needs a bit of love but we are on the right path. Needs lots of testing.