Flying does not block very fast speed anymore.
After many months of thinking about this change, I have decided to revert part of it. Partly due to some players expressing their dissatisfaction with the change, but also to preserve the awesomeness of flying. Some players thought I was targeting vampires in particular, but I will for the record state that was not the case. I do not make changes in reaction to players ascending and winning just to make the game harder. I have a set of design principles that I'm trying to adhere to.

I'm partly reverting 07f159d01 and 9369b34d2 so that only Levitation blocks fast speed. This puts a much greater emphasis on having flying and makes levitation worse. I think this makes more sense because levitation is much easier to get (via ring, potion, spell, or boots). Also, the Barbarian's quest artifact had been changed to grant flying instead of levitation so removing any nerf to flying also benefits the barbarian significantly.

One thing I may implement is a very sneak CrecelleHack monster spell Levitation, so that monsters can cast it at you...