Increase rate of artifact drops slightly; allow lenses and amulets to generate as artifacts.
This is an adjustment to make wishless playing a little more feasible.

The alternate artifact items is something I meant to do since Hack'EM! But it just kept getting nudged down the list. But now at least sometimes we'll see the Amulet of Storms or the Lenses of Truth randomly generate.