Port illusions and the mirror image spell from CrecelleHack.
This is a really cool mechanic and it makes higher level spellcasting monsters more threatening. It also raises the usefulness of protection from shapechangers (not that it needs more buffing). Did some minimal testing and it seems to function as promised. When taking off PFSC after unmasking illusions, they remain unmasked. I think this is normal so it's not a bug so far.