Fix: thick-skinned monsters were taking minimal damage from all weapons.
I noticed this after a while playing a valkyrie with a war hammer - attacking things like mumaks and mimics they were only taking 6 damage even with hammer at expert. Tracked this to a bad check in the weapon damage value routine where the actual material of the weapon was not being checked. Probably happened when I transferred object materials over.