Fix: artifacts were not generating with their appropriate material.
Since artifacts are a lot more rare, it took me a while to notice this but somehow I happened upon a random Grayswandir that was just an iron saber (imagine my disappointment).

I found some missing code in oname() where the material was setup in xnethack and confirmed artifacts are now getting the material correctly.