Inferior items are much more common.
It didn't make sense that the inferior property would be as rare as the exceptional prop - it should be the opposite. This brings the chance of inferior items generating to a whopping 1 in 20. This check is after the first check for rolling superior or exceptional items so that doesn't change their odds.