Loosen requirements for forging superior and exceptional items.
This mechanic was originally ported from EvilHack.

EvilHack has some stringent restrictions in place for forging these qualities - presumably because many more superior and exceptional items are provided by altar sacrifice. However, in NerfHack we offer no such luxuries because altars don't provide any sacrifice gifts, so I decided to make the forging system much more flexible.

EvilHack requires 2 superior items to forge another superior item, and likewise 2 exceptional items for an exceptional result. I have lowered that requirement to be just one of the two items having the appropriate quality.

Especially in the case of exceptional items - they are so incredibly rare that if one does pop up, we need to allow the player room to navigate the forging recipes to make full use of it. This also adds more tinkering and *fun* to the game. I also found myself wondering why the qualities didn't transfer when they were the second ingredient - this is how all the other parts of forging work for material and oprops. Object properties are arguably more powerful than quality so the additional restriction doesn't make sense (in my opinion).

I also lowered the enchantment requirements slightly from +3 for superior to +2, and +5 for exceptional to +4, so that in more cases you don't need to have the actual quality - enchantment will suffice (this is also because enchanting items in NerfHack is also much more difficult).