Allow ammo, missiles, magic, and non-metallic items to be eligible for build quality and alignments.
Not sure why I didn't notice this earlier, but this was a restriction inherited from EvilHack. I think this will seriously benefit the diversity of items in the game - especially for non-metal, which surprisingly I didn't really notice until the late-mid-game of a valkyrie I was playing. Also for ammo and missiles, although I can see the potential for a little overpowered-ness creeping in, I also think it opens up a lot of flexibility for forging possibilities.

The alignments will also be interesting to see spread out...