Port over EvilHack's advanced handling of hunger for vampires/dhampir.
Dhampir no longer die from hunger - they have the following effects as their hunger status advances (these conditions are temporary while the hunger status is active and are restored when the hunger status is cured/alleviated):
* From hungry to weak: you lose 1 point from ALL attributes.
* From weak to frail: you lose 5 points from ALL attributes and become deaf.
* From frail to starving: all attributes are set to 3 and you also become blind.

To accomodate these new changes, the new FRAIL hunger state was ported over, along with the FROMHUNGER slot for player property causes. In EvilHack, starving status also makes the player Slow - but we don't have that yet. I am thinking of adding it in the near future. I tested these changes and both advancing and retreating through hunger levels works.

This is a neat twist on the vampire that I have been eyeing for a long time, I think Evil did it pretty well and it kind of balances out the playstyle of vampires a bit. I am also thinking of using bits of this in my implementation of the Grung race because they also have attribute modification based on hydration level.