Adjust item material probabilities, influence from EvilHack's model.
Players will be happy to see that dragonhide is an option now. After playtesting for a while, I realized that xNetHack and EvilHack use fairly different models for the material distributions - so I merged them. I also converted all the probability tables to add to 1000 for readability and to standardize it.