Forging two of the same item creates more of the same item.
This is an idea I've had kicking around for a while to vastly improve the forging system. As it was, to create items, you could combine different items - but if you had a reasonable collection of the same items, it was awkward to improve upon them without a lot of roundabout paths. This enables you to take the best of your like items and get the most out of them. If you have a blessed superior long sword and a silver long sword - forge them together. If you then find a long sword of rage - go ahead and forge it together.

There is always a risk of creating an inferior product - that is the drawback, but otherwise I think this will be much better.

Two recipes combined identical items and had to be adjusted: long sword and katana.