Allow forging different levels of quality to be more flexible.
Previously, two forging ingredients needed to be exactly the same to produce the same result (with a small chance of something better). For example, a superior dagger must be combined with a superior dagger to make a superior dagger. If you combined an exceptional dagger with a superior dagger, you probably ended up with a regular dagger. This commit improves the combination formula so not everything is lost when combining items with different qualities. We take the average of both ingredient qualities and 50% of the time add a level. This encourages forging more and experimenting more.