Make projectile traps more consistent (from xNetHack).
This makes a trap shoot and yield the same type of item. If an arrow trap shoots silver arrows, it will yield those when disarmed. Same for dart and spear traps. This should dramatically cut down on the variety of single projectile types we see pop up that also annoyingly don't stack.