Juice up the chests in Moloch's Temple.
This branch, along with the Lost Tomb and Wyrm Caves, should have some nice incentives to conquer it. I added a 50/50 chance of a marker or wand of polymorph and many more scrolls can spawn in the chests.

I also detected a bug with specifying the wands in the tomb/temple levels and fixed it.