Revert "All crowning gifts are legendary."
This reverts commit 62c155cc6778eb67d6b64dafb7fa85a7e100100b.

This change introduces a few subtle inconsistencies that don't sit well with me. For example, Excalibur is both obtainable from #dipping for knights and from crowning. However, if the one from crowning is legendary, why wouldn't the fountain Excalibur also be?  Same goes for Mjolnir which the valk gets as a sac gift, but is also the crowning gift for priest. Too many awkward questions so I'm just keeping it simple...