Clamp the range of summon insects to 7 squares or less.
This is related to the last change in response to summon insects just being much too powerful and spammy for monsters to have. Not only could they summon migos, but in a level like Moloch's Temple with a really long hall - the 4 priests could all spam this from any range. This resulted in a very frustrating fight.