Scale wand ray damage with the zapper's level.
This is related to the previous change, but after reading some IRC comments I thought this would be worth experimenting with a bit. Wands of magic missile are notoriously weak after the early game so this would dramatically increase their value as both the player and monsters increase in level. I didn't stop there though, and made the change for all ray-based wands except drain life.